refactor: move crates into `penumbra-sdk` namespace (#4975)
This moves our workspace into a `penumbra-sdk` namespace ahead of
publishing `v0.82.0-alpha.0` on crates.io. We're using the `penumbra-sdk`
prefix for all crate names because it's unique. (The `penumbra`) crate
name is already taken on crates.io and belongs to an unrelated project.

In addition to a renaming, several other changes are made for crates.io
compatibility:

  * fills out crate metadata fields to meet spec
  * breaks out app-tests to separate crate